[FLINK-10912][rocksdb] Refactor RocksDBCustomizeOptions to RocksDBConfigurableOptions and ConfigurableOptionsFactory
