+++ docs/ops/config.md
@@ -162,13 +162,12 @@ The configuration keys in this section are independent of the used resource mana
 
 {% include generated/rocks_db_configuration.html %}
 
-### RocksDB Customized Options
-Certain RocksDB configurable options provided by Flink.
-These options have no default values, meaning they are not configured by default from Flink's side
-and leave them for predefined options and options factory to keep backward compatibility.
-The final RocksDB's option would be chosen through the priority of `OptionsFactory` --> `RocksDBCustomizedOptions` --> `PredefinedOptions`.
+### RocksDB Configurable Options
+Specific RocksDB configurable options, provided by Flink, to create a corresponding `ConfigurableOptionsFactory`.
+And the created one would be used as default `OptionsFactory` in `RocksDBStateBackend`
+unless user define a `OptionsFactory` and set via `RocksDBStateBackend.setOptions(optionsFactory)`
 
-{% include generated/rocks_db_customized_configuration.html %}
+{% include generated/rocks_db_configurable_configuration.html %}
 
 ### Queryable State
