+++ flink-state-backends/flink-statebackend-rocksdb/src/main/java/org/apache/flink/contrib/streaming/state/ConfigurableOptionsFactory.java
@@ -1,0 +1,418 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.contrib.streaming.state;
+
+import org.apache.flink.configuration.ConfigOption;
+import org.apache.flink.configuration.Configuration;
+import org.apache.flink.configuration.MemorySize;
+import org.apache.flink.util.Preconditions;
+
+import org.rocksdb.BlockBasedTableConfig;
+import org.rocksdb.ColumnFamilyOptions;
+import org.rocksdb.CompactionStyle;
+import org.rocksdb.DBOptions;
+import org.rocksdb.PlainTableConfig;
+import org.rocksdb.TableFormatConfig;
+
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import static org.apache.flink.contrib.streaming.state.RocksDBConfigurableOptions.BLOCK_CACHE_SIZE;
+import static org.apache.flink.contrib.streaming.state.RocksDBConfigurableOptions.BLOCK_SIZE;
+import static org.apache.flink.contrib.streaming.state.RocksDBConfigurableOptions.COMPACTION_STYLE;
+import static org.apache.flink.contrib.streaming.state.RocksDBConfigurableOptions.MAX_BACKGROUND_THREADS;
+import static org.apache.flink.contrib.streaming.state.RocksDBConfigurableOptions.MAX_OPEN_FILES;
+import static org.apache.flink.contrib.streaming.state.RocksDBConfigurableOptions.MAX_SIZE_LEVEL_BASE;
+import static org.apache.flink.contrib.streaming.state.RocksDBConfigurableOptions.MAX_WRITE_BUFFER_NUMBER;
+import static org.apache.flink.contrib.streaming.state.RocksDBConfigurableOptions.MIN_WRITE_BUFFER_NUMBER_TO_MERGE;
+import static org.apache.flink.contrib.streaming.state.RocksDBConfigurableOptions.TARGET_FILE_SIZE_BASE;
+import static org.apache.flink.contrib.streaming.state.RocksDBConfigurableOptions.USE_DYNAMIC_LEVEL_SIZE;
+import static org.apache.flink.contrib.streaming.state.RocksDBConfigurableOptions.WRITE_BUFFER_SIZE;
+
+/**
+ * A configurable implementation of {@link OptionsFactory} using options provided by {@link RocksDBConfigurableOptions}
+ * and acted as the default {@link RocksDBStateBackend#optionsFactory} if user not defined a {@link OptionsFactory}.
+ *
+ * <p>This implementation also provide some setters to let user could create a {@link OptionsFactory} conveniently。
+ */
+public class ConfigurableOptionsFactory implements OptionsFactory {
+
+	private final Map<String, String> configuredOptions = new HashMap<>();
+
+	@Override
+	public DBOptions createDBOptions(DBOptions currentOptions) {
+		if (isOptionConfigured(MAX_BACKGROUND_THREADS)) {
+			currentOptions.setIncreaseParallelism(getMaxBackgroundThreads());
+		}
+
+		if (isOptionConfigured(MAX_OPEN_FILES)) {
+			currentOptions.setMaxOpenFiles(getMaxOpenFiles());
+		}
+
+		return currentOptions;
+	}
+
+	@Override
+	public ColumnFamilyOptions createColumnOptions(ColumnFamilyOptions currentOptions) {
+		if (isOptionConfigured(COMPACTION_STYLE)) {
+			currentOptions.setCompactionStyle(getCompactionStyle());
+		}
+
+		if (isOptionConfigured(USE_DYNAMIC_LEVEL_SIZE)) {
+			currentOptions.setLevelCompactionDynamicLevelBytes(getUseDynamicLevelSize());
+		}
+
+		if (isOptionConfigured(TARGET_FILE_SIZE_BASE)) {
+			currentOptions.setTargetFileSizeBase(getTargetFileSizeBase());
+		}
+
+		if (isOptionConfigured(MAX_SIZE_LEVEL_BASE)) {
+			currentOptions.setMaxBytesForLevelBase(getMaxSizeLevelBase());
+		}
+
+		if (isOptionConfigured(WRITE_BUFFER_SIZE)) {
+			currentOptions.setWriteBufferSize(getWriteBufferSize());
+		}
+
+		if (isOptionConfigured(MAX_WRITE_BUFFER_NUMBER)) {
+			currentOptions.setMaxWriteBufferNumber(getMaxWriteBufferNumber());
+		}
+
+		if (isOptionConfigured(MIN_WRITE_BUFFER_NUMBER_TO_MERGE)) {
+			currentOptions.setMinWriteBufferNumberToMerge(getMinWriteBufferNumberToMerge());
+		}
+
+		TableFormatConfig tableFormatConfig = currentOptions.tableFormatConfig();
+
+		BlockBasedTableConfig blockBasedTableConfig;
+		if (tableFormatConfig == null) {
+			blockBasedTableConfig = new BlockBasedTableConfig();
+		} else {
+			if (tableFormatConfig instanceof PlainTableConfig) {
+				// if the table format config is PlainTableConfig, we just return current column-family options
+				return currentOptions;
+			} else {
+				blockBasedTableConfig = (BlockBasedTableConfig) tableFormatConfig;
+			}
+		}
+
+		if (isOptionConfigured(BLOCK_SIZE)) {
+			blockBasedTableConfig.setBlockSize(getBlockSize());
+		}
+
+		if (isOptionConfigured(BLOCK_CACHE_SIZE)) {
+			blockBasedTableConfig.setBlockCacheSize(getBlockCacheSize());
+		}
+
+		return currentOptions.setTableFormatConfig(blockBasedTableConfig);
+	}
+
+	public Map<String, String> getConfiguredOptions() {
+		return new HashMap<>(configuredOptions);
+	}
+
+	private boolean isOptionConfigured(ConfigOption configOption) {
+		return configuredOptions.containsKey(configOption.key());
+	}
+
+	//--------------------------------------------------------------------------
+	// Maximum number of concurrent background flush and compaction threads
+	//--------------------------------------------------------------------------
+
+	private int getMaxBackgroundThreads() {
+		return Integer.parseInt(getInternal(MAX_BACKGROUND_THREADS.key()));
+	}
+
+	public ConfigurableOptionsFactory setMaxBackgroundThreads(int totalThreadCount) {
+		Preconditions.checkArgument(totalThreadCount > 0);
+		configuredOptions.put(MAX_BACKGROUND_THREADS.key(), String.valueOf(totalThreadCount));
+		return this;
+	}
+
+	//--------------------------------------------------------------------------
+	// Maximum number of open files
+	//--------------------------------------------------------------------------
+
+	private int getMaxOpenFiles() {
+		return Integer.parseInt(getInternal(MAX_OPEN_FILES.key()));
+	}
+
+	public ConfigurableOptionsFactory setMaxOpenFiles(int maxOpenFiles) {
+		configuredOptions.put(MAX_OPEN_FILES.key(), String.valueOf(maxOpenFiles));
+		return this;
+	}
+
+	//--------------------------------------------------------------------------
+	// The style of compaction for DB.
+	//--------------------------------------------------------------------------
+
+	private CompactionStyle getCompactionStyle() {
+		return CompactionStyle.valueOf(getInternal(COMPACTION_STYLE.key()).toUpperCase());
+	}
+
+	public ConfigurableOptionsFactory setCompactionStyle(CompactionStyle compactionStyle) {
+		setInternal(COMPACTION_STYLE.key(), compactionStyle.name());
+		return this;
+	}
+
+	//--------------------------------------------------------------------------
+	// Whether to configure RocksDB to pick target size of each level dynamically.
+	//--------------------------------------------------------------------------
+
+	private boolean getUseDynamicLevelSize() {
+		return getInternal(USE_DYNAMIC_LEVEL_SIZE.key()).compareToIgnoreCase("false") != 0;
+	}
+
+	public ConfigurableOptionsFactory setUseDynamicLevelSize(boolean value) {
+		configuredOptions.put(USE_DYNAMIC_LEVEL_SIZE.key(), value ? "true" : "false");
+		return this;
+	}
+
+	//--------------------------------------------------------------------------
+	// The target file size for compaction, i.e., the per-file size for level-1
+	//--------------------------------------------------------------------------
+
+	private long getTargetFileSizeBase() {
+		return MemorySize.parseBytes(getInternal(TARGET_FILE_SIZE_BASE.key()));
+	}
+
+	public ConfigurableOptionsFactory setTargetFileSizeBase(String targetFileSizeBase) {
+		Preconditions.checkArgument(MemorySize.parseBytes(targetFileSizeBase) > 0,
+			"Invalid configuration " + targetFileSizeBase + " for target file size base.");
+		setInternal(TARGET_FILE_SIZE_BASE.key(), targetFileSizeBase);
+		return this;
+	}
+
+
+	//--------------------------------------------------------------------------
+	// Maximum total data size for a level, i.e., the max total size for level-1
+	//--------------------------------------------------------------------------
+
+	private long getMaxSizeLevelBase() {
+		return MemorySize.parseBytes(getInternal(MAX_SIZE_LEVEL_BASE.key()));
+	}
+
+	public ConfigurableOptionsFactory setMaxSizeLevelBase(String maxSizeLevelBase) {
+		Preconditions.checkArgument(MemorySize.parseBytes(maxSizeLevelBase) > 0,
+			"Invalid configuration " + maxSizeLevelBase + " for max size of level base.");
+		setInternal(MAX_SIZE_LEVEL_BASE.key(), maxSizeLevelBase);
+		return this;
+	}
+
+	//--------------------------------------------------------------------------
+	// Amount of data to build up in memory (backed by an unsorted log on disk)
+	// before converting to a sorted on-disk file. Larger values increase
+	// performance, especially during bulk loads.
+	//--------------------------------------------------------------------------
+
+	private long getWriteBufferSize() {
+		return MemorySize.parseBytes(getInternal(WRITE_BUFFER_SIZE.key()));
+	}
+
+	public ConfigurableOptionsFactory setWriteBufferSize(String writeBufferSize) {
+		Preconditions.checkArgument(MemorySize.parseBytes(writeBufferSize) > 0,
+			"Invalid configuration " + writeBufferSize + " for write-buffer size.");
+
+		setInternal(WRITE_BUFFER_SIZE.key(), writeBufferSize);
+		return this;
+	}
+
+	//--------------------------------------------------------------------------
+	// The maximum number of write buffers that are built up in memory.
+	//--------------------------------------------------------------------------
+
+	private int getMaxWriteBufferNumber() {
+		return Integer.parseInt(getInternal(MAX_WRITE_BUFFER_NUMBER.key()));
+	}
+
+	public ConfigurableOptionsFactory setMaxWriteBufferNumber(int writeBufferNumber) {
+		Preconditions.checkArgument(writeBufferNumber > 0,
+			"Invalid configuration " + writeBufferNumber + " for max write-buffer number.");
+		setInternal(MAX_WRITE_BUFFER_NUMBER.key(), Integer.toString(writeBufferNumber));
+		return this;
+	}
+
+	//--------------------------------------------------------------------------
+	// The minimum number that will be merged together before writing to storage
+	//--------------------------------------------------------------------------
+
+	private int getMinWriteBufferNumberToMerge() {
+		return Integer.parseInt(getInternal(MIN_WRITE_BUFFER_NUMBER_TO_MERGE.key()));
+	}
+
+	public ConfigurableOptionsFactory setMinWriteBufferNumberToMerge(int writeBufferNumber) {
+		Preconditions.checkArgument(writeBufferNumber > 0,
+			"Invalid configuration " + writeBufferNumber + " for min write-buffer number to merge.");
+		setInternal(MIN_WRITE_BUFFER_NUMBER_TO_MERGE.key(), Integer.toString(writeBufferNumber));
+		return this;
+	}
+
+	//--------------------------------------------------------------------------
+	// Approximate size of user data packed per block. Note that the block size
+	// specified here corresponds to uncompressed data. The actual size of the
+	// unit read from disk may be smaller if compression is enabled
+	//--------------------------------------------------------------------------
+
+	private long getBlockSize() {
+		return MemorySize.parseBytes(getInternal(BLOCK_SIZE.key()));
+	}
+
+	public ConfigurableOptionsFactory setBlockSize(String blockSize) {
+		Preconditions.checkArgument(MemorySize.parseBytes(blockSize) > 0,
+			"Invalid configuration " + blockSize + " for block size.");
+		setInternal(BLOCK_SIZE.key(), blockSize);
+		return this;
+	}
+
+	//--------------------------------------------------------------------------
+	// The amount of the cache for data blocks in RocksDB
+	//--------------------------------------------------------------------------
+
+	private long getBlockCacheSize() {
+		return MemorySize.parseBytes(getInternal(BLOCK_CACHE_SIZE.key()));
+	}
+
+	public ConfigurableOptionsFactory setBlockCacheSize(String blockCacheSize) {
+		Preconditions.checkArgument(MemorySize.parseBytes(blockCacheSize) > 0,
+			"Invalid configuration " + blockCacheSize + " for block cache size.");
+		setInternal(BLOCK_CACHE_SIZE.key(), blockCacheSize);
+
+		return this;
+	}
+
+	private static final String[] CANDIDATE_CONFIGS = new String[]{
+		// configurable DBOptions
+		MAX_BACKGROUND_THREADS.key(),
+		MAX_OPEN_FILES.key(),
+
+		// configurable ColumnFamilyOptions
+		COMPACTION_STYLE.key(),
+		USE_DYNAMIC_LEVEL_SIZE.key(),
+		TARGET_FILE_SIZE_BASE.key(),
+		MAX_SIZE_LEVEL_BASE.key(),
+		WRITE_BUFFER_SIZE.key(),
+		MAX_WRITE_BUFFER_NUMBER.key(),
+		MIN_WRITE_BUFFER_NUMBER_TO_MERGE.key(),
+		BLOCK_SIZE.key(),
+		BLOCK_CACHE_SIZE.key()
+	};
+
+	private static final Set<String> POSITIVE_INT_CONFIG_SET = new HashSet<>(Arrays.asList(
+		MAX_BACKGROUND_THREADS.key(),
+		MAX_WRITE_BUFFER_NUMBER.key(),
+		MIN_WRITE_BUFFER_NUMBER_TO_MERGE.key()
+	));
+
+	private static final Set<String> SIZE_CONFIG_SET = new HashSet<>(Arrays.asList(
+		TARGET_FILE_SIZE_BASE.key(),
+		MAX_SIZE_LEVEL_BASE.key(),
+		WRITE_BUFFER_SIZE.key(),
+		BLOCK_SIZE.key(),
+		BLOCK_CACHE_SIZE.key()
+	));
+
+	private static final Set<String> BOOLEAN_CONFIG_SET = new HashSet<>(Collections.singletonList(
+		USE_DYNAMIC_LEVEL_SIZE.key()
+	));
+
+	private static final Set<String> COMPACTION_STYLE_SET = Arrays.stream(CompactionStyle.values())
+		.map(c -> c.name().toLowerCase()).collect(Collectors.toSet());
+
+	/**
+	 * Creates a {@link RocksDBConfigurableOptions} instance from a {@link Configuration}.
+	 *
+	 * <p>If no options within {@link RocksDBConfigurableOptions} has ever been configured,
+	 * the created OptionsFactory would not override anything defined in {@link PredefinedOptions}.
+	 *
+	 * @param configuration Configuration to be used for the ConfigurableOptionsFactory creation
+	 * @return A ConfigurableOptionsFactory created from the given configuration
+	 */
+	public static ConfigurableOptionsFactory fromConfig(Configuration configuration) {
+		ConfigurableOptionsFactory customizedOptions = new ConfigurableOptionsFactory();
+		for (String key : CANDIDATE_CONFIGS) {
+			String newValue = configuration.getString(key, null);
+
+			if (newValue != null) {
+				if (checkArgumentValid(key, newValue)) {
+					customizedOptions.configuredOptions.put(key, newValue);
+				}
+			}
+		}
+		return customizedOptions;
+	}
+
+	/**
+	 * Helper method to check whether the (key,value) is valid through given configuration and returns the formatted value.
+	 *
+	 * @param key The configuration key which is configurable in {@link RocksDBConfigurableOptions}.
+	 * @param value The value within given configuration.
+	 *
+	 * @return whether the given key and value in string format is legal.
+	 */
+	private static boolean checkArgumentValid(String key, String value) {
+		if (POSITIVE_INT_CONFIG_SET.contains(key)) {
+
+			Preconditions.checkArgument(Integer.parseInt(value) > 0,
+				"Configured value for key: " + key + " must be larger than 0.");
+		} else if (SIZE_CONFIG_SET.contains(key)) {
+
+			Preconditions.checkArgument(MemorySize.parseBytes(value) > 0,
+				"Configured size for key" + key + " must be larger than 0.");
+		} else if (BOOLEAN_CONFIG_SET.contains(key)) {
+
+			Preconditions.checkArgument("true".equalsIgnoreCase(value) || "false".equalsIgnoreCase(value),
+				"The configured boolean value: " + value + " for key: " + key + " is illegal.");
+		} else if (key.equals(COMPACTION_STYLE.key())) {
+			value = value.toLowerCase();
+			Preconditions.checkArgument(COMPACTION_STYLE_SET.contains(value),
+				"Compression type: " + value + " is not recognized with legal types: " + String.join(", ", COMPACTION_STYLE_SET));
+		}
+		return true;
+	}
+
+	/**
+	 * Sets the configuration with (key, value) if the key is predefined, otherwise throws IllegalArgumentException.
+	 *
+	 * @param key The configuration key, if key is not predefined, throws IllegalArgumentException out.
+	 * @param value The configuration value.
+	 */
+	private void setInternal(String key, String value) {
+		Preconditions.checkArgument(value != null && !value.isEmpty(),
+			"The configuration value must not be empty.");
+
+		configuredOptions.put(key, value);
+	}
+
+	/**
+	 * Returns the value in string format with the given key.
+	 *
+	 * @param key The configuration-key to query in string format.
+	 */
+	private String getInternal(String key) {
+		Preconditions.checkArgument(configuredOptions.containsKey(key),
+			"The configuration " + key + " has not been configured.");
+
+		return configuredOptions.get(key);
+	}
+}
