+++ flink-state-backends/flink-statebackend-rocksdb/src/main/java/org/apache/flink/contrib/streaming/state/RocksDBStateBackend.java
@@ -124,10 +124,7 @@ public class RocksDBStateBackend extends AbstractStateBackend implements Configu
 
 	/** The pre-configured option settings. */
 	@Nullable
-	private PredefinedOptions predefinedOptions = null;
-
-	/** The customized RocksDB options, default value: empty-map, means no customized options. */
-	private RocksDBCustomizedOptions customizedOptions;
+	private PredefinedOptions predefinedOptions;
 
 	/** The options factory to create the RocksDB options in the cluster. */
 	@Nullable
@@ -270,7 +267,6 @@ public class RocksDBStateBackend extends AbstractStateBackend implements Configu
 		this.priorityQueueStateType = PriorityQueueStateType.HEAP;
 		this.defaultMetricOptions = new RocksDBNativeMetricOptions();
 		this.enableTtlCompactionFilter = TernaryBoolean.UNDEFINED;
-		this.customizedOptions = new RocksDBCustomizedOptions();
 	}
 
 	/**
@@ -347,12 +343,9 @@ public class RocksDBStateBackend extends AbstractStateBackend implements Configu
 		this.predefinedOptions = original.predefinedOptions == null ?
 			PredefinedOptions.valueOf(config.getString(RocksDBOptions.PREDEFINED_OPTIONS)) : original.predefinedOptions;
 
-		// configure RocksDB customized options
-		RocksDBCustomizedOptions customizedOptionsFromConfig = RocksDBCustomizedOptions.fromConfig(config);
-		this.customizedOptions = customizedOptionsFromConfig.overrideBy(original.customizedOptions);
-
-		// copy remaining settings
-		this.optionsFactory = original.optionsFactory;
+		// configure RocksDB options factory
+		this.optionsFactory = original.optionsFactory == null ?
+			ConfigurableOptionsFactory.fromConfig(config) : original.optionsFactory;
 	}
 
 	// ------------------------------------------------------------------------
@@ -675,10 +668,8 @@ public class RocksDBStateBackend extends AbstractStateBackend implements Configu
 	/**
 	 * Sets the predefined options for RocksDB.
 	 *
-	 * <p>If customized options is configured, then the options from the {@link RocksDBCustomizedOptions}
-	 * are applied on top of the predefined options.
-	 *
-	 * <p>If a user-defined options factory is set (via {@link #setOptions(OptionsFactory)}),
+	 * <p>If user-configured options within {@link RocksDBConfigurableOptions} is set (through flink-conf.yaml)
+	 * or a user-defined options factory is set (via {@link #setOptions(OptionsFactory)}),
 	 * then the options from the factory are applied on top of the here specified
 	 * predefined options and customized options.
 	 *
@@ -693,10 +684,8 @@ public class RocksDBStateBackend extends AbstractStateBackend implements Configu
 	 * The default options (if nothing was set via {@link #setPredefinedOptions(PredefinedOptions)})
 	 * are {@link PredefinedOptions#DEFAULT}.
 	 *
-	 * <p>If customized options is configured, then the options from the {@link RocksDBCustomizedOptions}
-	 * are applied on top of the predefined options.
-	 *
-	 * <p>If a user-defined options factory is set (via {@link #setOptions(OptionsFactory)}),
+	 * <p>If user-configured options within {@link RocksDBConfigurableOptions} is set (through flink-conf.yaml)
+	 * of a user-defined options factory is set (via {@link #setOptions(OptionsFactory)}),
 	 * then the options from the factory are applied on top of the predefined and customized options.
 	 *
 	 * @return The currently set predefined options for RocksDB.
@@ -709,44 +698,15 @@ public class RocksDBStateBackend extends AbstractStateBackend implements Configu
 	}
 
 	/**
-	 * Sets the customized options on top of predefined options.
-	 *
-	 * <p>If a user-defined options factory is set (via {@link #setOptions(OptionsFactory)}),
-	 * then the options from the factory are applied on top of the predefined and customized options.
-	 *
-	 * @param customizedOptions The customized options to set (must not be null).
-	 */
-	public void setCustomizedOptions(RocksDBCustomizedOptions customizedOptions) {
-		this.customizedOptions = checkNotNull(customizedOptions);
-	}
-
-	/**
-	 * Gets the currently user customized options for RocksDB.
-	 * The default customized options are empty, please set provided specific options for {@link RocksDBCustomizedOptions}.
-	 *
-	 * <p>The customized options here are applied on top of the pre-defined options profile
-	 * selected via {@link #setPredefinedOptions(PredefinedOptions)}.
-	 *
-	 * <p>If a user-defined  options factory is set (via {@link #setOptions(OptionsFactory)}),
-	 * then the options from the factory are applied on top of the predefined and customized options.
-	 *
-	 * @return The currently user customized options for RocksDB.
-	 */
-	public RocksDBCustomizedOptions getCustomizedOptions() {
-		return customizedOptions;
-	}
-
-	/**
 	 * Sets {@link org.rocksdb.Options} for the RocksDB instances.
 	 * Because the options are not serializable and hold native code references,
 	 * they must be specified through a factory.
 	 *
 	 * <p>The options created by the factory here are applied on top of the pre-defined
-	 * options and customized options.
-	 *
-	 * <p>If the pre-defined options profile is the default
-	 * ({@link PredefinedOptions#DEFAULT}), and no customized options are specified,
-	 * then the factory fully controls the RocksDB options.
+	 * options profile selected via {@link #setPredefinedOptions(PredefinedOptions)}.
+	 * If the pre-defined options profile is the default
+	 * ({@link PredefinedOptions#DEFAULT}), then the factory fully controls the RocksDB
+	 * options.
 	 *
 	 * @param optionsFactory The options factory that lazily creates the RocksDB options.
 	 */
@@ -770,9 +730,6 @@ public class RocksDBStateBackend extends AbstractStateBackend implements Configu
 		// initial options from pre-defined profile
 		DBOptions opt = getPredefinedOptions().createDBOptions();
 
-		// add user-customized options, if specified
-		opt = customizedOptions.getDBOptions(opt);
-
 		// add user-defined options factory, if specified
 		if (optionsFactory != null) {
 			opt = optionsFactory.createDBOptions(opt);
@@ -790,9 +747,6 @@ public class RocksDBStateBackend extends AbstractStateBackend implements Configu
 	public ColumnFamilyOptions getColumnOptions() {
 		// initial options from pre-defined profile
 		ColumnFamilyOptions opt = getPredefinedOptions().createColumnOptions();
-
-		// add user-customized options, if specified
-		opt = customizedOptions.getColumnOptions(opt);
 
 		// add user-defined options, if specified
 		if (optionsFactory != null) {
